fix : createCars 메소드 변경
