+++ src/main/java/racingcar/Application.java
@@ -1,14 +1,11 @@
 package racingcar;
 
-import static racingcar.view.InputView.askAttemptNumber;
-import static racingcar.view.InputView.askCarName;
-
-import racingcar.domain.Car;
-import racingcar.view.InputView;
+import racingcar.controller.Game;
 
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-
+        Game game = new Game();
+        game.start();
     }
 }
