+++ src/main/java/racingcar/controller/Game.java
@@ -1,0 +1,7 @@
+package racingcar.controller;
+
+public class Game {
+    public void start(){
+
+    }
+}
