+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,0 +1,14 @@
+package racingcar.domain;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+class CarTest {
+
+    @Test
+    void createCars() {
+
+
+    }
+}
